feat(rpc): port Agave 4.1.0 parsed-account type changes
- JsonParsedVoteInfo: blockRevenueCommissionBps and
  inflationRewardsCommissionBps changed from `BigInt?` to `int?` (Agave
  returns small bounded u16 integers).
- JsonParsedVote latency changed from `BigInt?` to `int?` (small bounded
  u8 integer).
- JsonParsedRentInfo: now a union of the Agave 4.1.0+ shape
  (`lamportsPerByte: StringifiedBigInt?`) and the deprecated pre-4.1.0
  shape (`burnPercent`/`exemptionThreshold`/`lamportsPerByteYear`, all
  @Deprecated); narrow on `lamportsPerByte != null` vs
  `lamportsPerByteYear != null`.
- JsonParsedStakeDelegation.warmupCooldownRate made optional (Agave
  4.1.0 removed it from the parsed output; present only on older
  validators).
- JsonParsedStakeConfigInfo slashPenalty/warmupCooldownRate @Deprecated
  (the stake config program is no longer recognized by the RPC JSON
  parser as of Agave 4.1.0).
- allowed_numeric_keypaths: keep vote commissions and latency as `int`
  rather than upcasting to `BigInt` (generic parsed-accounts config and
  getVoteAccounts-specific).

1107 tests pass; --fatal-infos clean. `replacementBlockhash` always-present
is a TS type-only change with no Dart counterpart (the Dart port does not
model the simulateTransaction response `replacementBlockhash` as a typed
field).

+++ packages/solana_kit_rpc_parsed_types/lib/src/vote_accounts.dart
@@ -39,7 +39,10 @@ class JsonParsedVoteInfo {
   final Address? blockRevenueCollector;
 
   /// Block revenue commission in basis points, if reported.
-  final BigInt? blockRevenueCommissionBps;
+  //
+  // Changed from `BigInt?` to `int?` in @solana/kit v7.0.0 to match Agave
+  // 4.1.0, which returns this as a small bounded `u16` integer.
+  final int? blockRevenueCommissionBps;
 
   /// The compressed BLS public key, or `null` if absent or not reported.
   final String? blsPubkeyCompressed;
@@ -54,7 +57,10 @@ class JsonParsedVoteInfo {
   final Address? inflationRewardsCollector;
 
   /// Inflation rewards commission in basis points, if reported.
-  final BigInt? inflationRewardsCommissionBps;
+  //
+  // Changed from `BigInt?` to `int?` in @solana/kit v7.0.0 to match Agave
+  // 4.1.0, which returns this as a small bounded `u16` integer.
+  final int? inflationRewardsCommissionBps;
 
   /// The last timestamp recorded by this vote account.
   final JsonParsedLastTimestamp lastTimestamp;
@@ -280,7 +286,10 @@ class JsonParsedVote {
   final int confirmationCount;
 
   /// The latency of this vote, in slots, if reported.
-  final BigInt? latency;
+  //
+  // Changed from `BigInt?` to `int?` in @solana/kit v7.0.0 to match Agave
+  // 4.1.0, which returns this as a small bounded `u8` integer.
+  final int? latency;
 
   /// The slot that was voted on.
   final Slot slot;
